pyproject.toml
# Nexus AI Assistant Setup

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-ai-assistant"
version = "1.0.0"
description = "An intelligent AI assistant for natural language processing and task automation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nexus Team", email = "support@nexus-ai.com"}
]
keywords = ["ai", "assistant", "chatbot", "openai", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
]
nlp = [
    "nltk>=3.8.1",
    "spacy>=3.6.0",
    "transformers>=4.30.0",
]
data = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
]
all = [
    "nexus-ai-assistant[web,nlp,data,dev]"
]

[project.scripts]
nexus = "nexus.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/nexus"
Repository = "https://github.com/yourusername/nexus.git"
Documentation = "https://docs.nexus-ai.com"
"Bug Tracker" = "https://github.com/yourusername/nexus/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nexus = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
